Added a new source and header file to the external 'console' directory, and the WICED SDK builds them just fine
The idea here was to illustrate how to point the WICED SDK build
system at sources and headers we're adding to an app located in a
directory external to the SDK.

The file "stringer.c" is not an original file - it is just a copy of
the file "uart.c" that lives in the WICED SDK's "app/snip/uart"
sample, and has been re-located to the external "console" directory.

But before echoing characters received over the UART interface back
out, "stringer.c" now calls a new function, "stringer_char_xform(),"
which simply converts upper-case characters to lower-case, and vice
versa.

"stringer_char_xform()" is declared in the header file
"stringer_pal.h". And "stringer_pal.h" is added to a list of include
directories maintained within the WICED SDK by the following line in
the "stringer_pal.mk" file (the "glue" that allows us to point the
WICED SDK build system at our files):

  $(NAME)_INCLUDES        += STRINGER_PAL_PATH

The macro "$(NAME)_INCLUDES" expands to "App_Stringer_INCLUDES" when
the WICED SDK process it (via the WICED SDK makefile located at
"apps/snip/stringer/stringer.mk").

Finally, "stringer_char_xform()" is defined in the new source file
"console/stringer_pal.c."