Define "main()" in our own code (pulling it out of the WICED SDK -- required tweaing the "WICED-SDK-3.7.0-7" git sub-module)
Used the WICED SDK's "ALTERNATE_MAIN" compile flag mechanism to
disable the standard/default FreeRTOS definition of main(), and
replace it with a main() of our own.

This required a tweak of the "WICED-SDK-3.7.0-7" git sub-module (see
commit "64ca018f2efbdf8d9c3561a50c4b62b5c825c755" in that repo for
more detail).

The version of "main()" that's defined in the new file "main.c" here
is a copy of the WICED SDK's built-in "main()" for a configuration
using FreeRTOS (this original "main()" can be found in WICED SDK file
"WICED-SDK-3.7.0-7/WICED/RTOS/FreeRTOS/WICED/wiced_rtos.c").

Pulling main() out of the SDK like this will allow us to define
additional FreeRTOS task threads and the like, without the need to
touch/modify the WICED SDK itself.