Added another simple task just to illustrate the process (this one watches button presses)
You simply get a message sent out the console, saying either
"Button 1 Pressed" or "Button 2 Pressed." No big deal.

The implementation of a new task can be anywhere (I stuck it in
stringer_pal.c here). But the routine implementing the task, and a
corresponding task handle, must be visible inside of main.c since this
is were we start the FreeRTOS scheduler.